# Implement stacktrie to improve RLP/hashing performance