--- a/trie/node.go
+++ b/trie/node.go
@@ -29,6 +29,7 @@ var indices = []string{"0", "1", "2", "3", "4", "5", "6", "7", "8", "9", "a", "b
 
 type node interface {
 	fstring(string) string
+	encode(w rlp.EncoderBuffer)
 	cache() (hashNode, bool)
 }